Moved `generateGeminiModel()` to `GeminiHelper`
